+++ Plugin.cs
@@ -22,6 +22,9 @@ namespace Dread
         private readonly Harmony _harmony = new(GUID);
         private EventHandler? _logLevelHandler;
 
+        private static bool MonsterPatchesEnabled =>
+            DreadConfig.MonsterAggressionEnabled.Value && !DreadConfig.CompatibilityMode.Value;
+
         private void Awake()
         {
             PluginDependencyResolver.Register();
@@ -34,28 +37,14 @@ namespace Dread
             _logLevelHandler = (_, _) => LoggingService.SetLevel(DreadConfig.LogLevelEntry.Value);
             DreadConfig.LogLevelEntry.SettingChanged += _logLevelHandler;
 
-            if (DreadConfig.MonsterAggressionEnabled.Value)
-            {
-                EnemyNavMeshAgentAwakePatch.Apply(_harmony);
-                EnemyDirectorSetInvestigatePatch.Apply(_harmony);
-            }
+            ApplyMonsterPatches();
             if (DreadConfig.CrouchSpeedBoostEnabled.Value)
                 PlayerControllerAwakePatch.Apply(_harmony);
-            ErrorReportPatch.Apply(_harmony);
+            if (DreadConfig.DebugConsoleGuardEnabled.Value)
+                DebugConsoleGuardPatch.Apply(_harmony);
 
-            DreadConfig.MonsterAggressionEnabled.SettingChanged += (_, _) =>
-            {
-                if (DreadConfig.MonsterAggressionEnabled.Value)
-                {
-                    EnemyNavMeshAgentAwakePatch.Apply(_harmony);
-                    EnemyDirectorSetInvestigatePatch.Apply(_harmony);
-                }
-                else
-                {
-                    EnemyNavMeshAgentAwakePatch.Remove(_harmony);
-                    EnemyDirectorSetInvestigatePatch.Remove(_harmony);
-                }
-            };
+            DreadConfig.MonsterAggressionEnabled.SettingChanged += (_, _) => ApplyMonsterPatches();
+            DreadConfig.CompatibilityMode.SettingChanged += (_, _) => ApplyMonsterPatches();
 
             DreadConfig.CrouchSpeedBoostEnabled.SettingChanged += (_, _) =>
             {
@@ -65,10 +54,32 @@ namespace Dread
                     PlayerControllerAwakePatch.Remove(_harmony);
             };
 
+            DreadConfig.DebugConsoleGuardEnabled.SettingChanged += (_, _) =>
+            {
+                if (DreadConfig.DebugConsoleGuardEnabled.Value)
+                    DebugConsoleGuardPatch.Apply(_harmony);
+                else
+                    DebugConsoleGuardPatch.Remove(_harmony);
+            };
+
             LoggingService.PrintAsciiArt();
             LoggingService.LogInfo($"{NAME} v{VERSION} loaded.");
 
             SceneManager.sceneLoaded += OnSceneLoaded;
+        }
+
+        private void ApplyMonsterPatches()
+        {
+            if (MonsterPatchesEnabled)
+            {
+                EnemyNavMeshAgentAwakePatch.Apply(_harmony);
+                EnemyDirectorSetInvestigatePatch.Apply(_harmony);
+            }
+            else
+            {
+                EnemyNavMeshAgentAwakePatch.Remove(_harmony);
+                EnemyDirectorSetInvestigatePatch.Remove(_harmony);
+            }
         }
 
         private static void OnSceneLoaded(Scene scene, LoadSceneMode mode)
@@ -84,4 +95,3 @@ namespace Dread
         }
     }
 }
-